backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
click>=8.0.0  # Required by uvicorn[standard]
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# AI Provider SDKs
# Updated versions for Agno 2.3.2 compatibility
openai>=2.0.0  # Agno 2.3.2 requires openai>=2.0.0 for openai.types.responses
anthropic>=0.50.0  # Agno requires newer anthropic for streaming types
google-generativeai==0.8.3  # Legacy Google AI SDK
google-genai>=1.0.0  # Required by Agno for Gemini models

# Agno Framework for Multi-Agent Systems (phidata)
agno>=2.0.0
# Vector database and embeddings for RAG
sentence-transformers>=2.2.0

# FastMCP for MCP Server Integration
fastmcp==0.2.0

# Database
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pgvector==0.3.6

# Redis for Caching
redis==5.2.1
hiredis==3.0.0

# OAuth & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
authlib==1.3.2

# Integrations
PyGithub==2.5.0
jira==3.8.0
atlassian-python-api==3.41.18

# Vector Operations
numpy==2.2.1
pgvector==0.3.6

# HTTP & Async
httpx==0.28.1
aiohttp==3.11.11
websockets==14.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==9.0.0
structlog==24.4.0
uuid7==0.1.0
cryptography==43.0.1
markdown==3.7
pypdf==5.1.0  # PDF text extraction

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
